+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Karafka framework changelog
 
 ## 2.3.2 (Unreleased)
+- **[Feature]** Provide swarm capabilities
+- [Enhancement] Introduce K8s swarm liveness listener.
 - [Fix] Quieting mode causes too early unsubscribe.
 
 ## 2.3.1 (2024-02-08)